Fix compiler warnings.

+++ Classes/DDASLLogger.m
@@ -91,7 +91,7 @@ static DDASLLogger *sharedInstance;
         snprintf(readUIDString, sizeof(readUIDString), "%d", readUID);
 #endif
 
-        NSAssert(l < sizeof(readUIDString),
+        NSAssert(l < (int)sizeof(readUIDString),
                  @"Formatted euid is too long.");
         NSAssert(aslLogLevel < (sizeof(level_strings) / sizeof(level_strings[0])),
                  @"Unhandled ASL log level.");
